+++ src/container/Register/SuccessStep.jsx
@@ -1,5 +1,4 @@
-import React from 'react'
-import { useEffect } from 'react'
+import React, { useEffect } from 'react'
 import { Link } from 'react-router-dom'
 import { successBg, sucessIcon } from '../../assets/images'
 import styled from '@emotion/styled'